Move `node:stream/consumers` types to `src`
This where Wrangler stores its other `.d.ts` files.
