+++ packages/wrangler/tsconfig.emit.json
@@ -12,6 +12,6 @@
 		"emitDeclarationOnly": true,
 		"outDir": "emitted-types"
 	},
-	"files": ["src/api/index.ts", "streams.d.ts"],
+	"files": ["src/api/index.ts"],
 	"include": ["src/*.d.ts"]
 }
